+++ unreal/CodexUnrealBlueprint/Source/CodexUnrealBlueprintEditor/Private/CodexDocumentImportGuard.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include "CoreMinimal.h"
+
+class SWindow;
+
+namespace CodexUnrealBlueprint
+{
+    class FDocumentImportGuard
+    {
+    public:
+        ~FDocumentImportGuard();
+        void Start();
+        void Stop();
+
+        static bool IsGeneratedDocumentPath(const FString& PackagePath);
+
+    private:
+        void TickModalLoop(float DeltaTime);
+
+        FDelegateHandle ModalLoopHandle;
+        TWeakPtr<SWindow> InspectedWindow;
+    };
+}